feat(operator): Add fail-over capabilities (#2)
* feat(operator): Add fail-over capabilities

This PR adds a new health_controller that takes care
of marking which pods as a replica, and which is master. This is done
by listening to all pod events that include the dragonfly label,
and then marking it either to be `master` or `replica` based
on the current state.

This PR also includes a `print-roles.go` helper CLI that essentially
shows the pods along with their roles given a dragonfly instance.