fix(tests): Run the operator separately
Currently, Our operator test suite ran the operator locally
which meant we had to set port-forwards to talk to induvidual
pods when required to be marked as replicas or as masters. These
port-forwards are error-prone and induced a lot of retry logic
into the application code making errors confusing.

This commit fixes that by essentially updating the suite_test.go
to run the tests inside a already operator installed cluster.