fix(tests): Run the operator separately (#20)
* fix(tests): Run the operator separately

Currently, Our operator test suite ran the operator locally
which meant we had to set port-forwards to talk to induvidual
pods when required to be marked as replicas or as masters. These
port-forwards are error-prone and induced a lot of retry logic
into the application code making errors confusing.

This commit fixes that by essentially updating the suite_test.go
to run the tests inside a already operator installed cluster.

Currently, As mentioned the port-forward issue is prone to failures
with GitHub actions because of the whole ipv6 and other issues.

As the role label is only added after we run the command on the pod,
Just checking for the role label should be enough.