feat(Operator): Handle updates to resources (#15)
* feat(Operator): Handle updates to resources

This commit adds the logic to handle updates to the
`Dragonfly` resource. This is done by adding an additional
variant in the df controller that uses update functions to
get latest resources. Logic has also been added into
the lifecycle controller to handle pod deletions when replicas
are reduced, etc.

All these cases are covered with tests.