fix(dfi): check pod ip before marking
Now that we mark pods as they get instanced without waiting for all
resources to be created the code cannot expect all pods to be ready
while it configures replication. This adds an extra pod IP Check.

If there is a pod without IP, the controller will wait for the next
pod event which gets triggered automatically once Kubernetes adds the
IP.