fix(rollout): Update rollout strategy to be `OnDelete` (#59)
* fix(rollout): Update rollout strategy to be `OnDelete`

Fixes https://github.com/dragonflydb/dragonfly-operator/issues/48

This commit updates the statefulset's upgrade strategy to be `OnDelete`
which means that the statefulset will not be automatically upgraded
when the pod template changes. Instead, The Operator will delete
each pod in the statefulset one at a time. This gives us the following
benefits:
- The order of pod rollout is controlled by the Operator which means
  we can start from replicas and go until master.
- The Operator can wait for the pod to be ready and also be in stable
  sync before it moves to the next pod.
- fix(resource): Remove `PhaseRollingUpdate` and use boolean