feat(upgrade): Use repltakeover (#109)
nstead of deleting the master, This PR updates the rollout flow to run repltakeover on a latest replica which would mark this pod as the master while terminating the old master. When a pod is terminated, Kubernetes automatically recreates it and the pod lifecycle controller will do the needful.

The pod lifecycle controller does react on various instances and tries to corrupt, but we only expect it to handle new pod creations and hence toggle is added to not react on `isRollingUpdate=true` for other events.