fix(controller): Disconnect clients when changing from master to replica #324
When a failover is triggered while the master is still running with
connections, we need to terminate those connections after it's removed
from the service to force a reconnect to the new master. This can
happen during updates to k8s on the node.

This also names the redis connections, so that those are left connected.